tools/requirements_tools.txt
requests>=2.25
numpy>=1.19
Pillow>=8.0
pyserial>=3.5

